+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - better error reporting on unknown import
 - check on number of arguments passed to `type`
 - warning when the formatter deletes comment(s) by mistake
+- check on arguments passed to `list`, `concat`, `append` and friends to only push valid nodes (that produces a value)
 
 ### Changed
 - instructions are on 4 bytes: 1 byte for the instruction, 1 byte of padding, 2 bytes for an immediate argument